+++ packages/core/components/hoc/withLoader/styles.css
@@ -1,3 +1,4 @@
+/* stylelint-disable selector-max-specificity */
 .transition.fade {
   opacity: 0;
 }
@@ -5,3 +6,4 @@
 .transition.fade:global(.fade-enter-done) {
   opacity: 1;
 }
+/* stylelint-enable */
